requirements.txt
alembic==1.16.4
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
bcrypt==4.3.0
beautifulsoup4==4.13.4
cachetools==6.1.0
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.2
croniter==6.0.0
cryptography==45.0.5
debtcollector==3.0.0
decorator==5.2.1
distro==1.9.0
dnspython==2.7.0
dogpile.cache==1.4.0
eval_type_backport==0.2.2
eventlet==0.40.1
fasteners==0.19
futurist==3.1.1
greenlet==3.2.3
groq==0.29.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_metadata==8.7.0
iso8601==2.1.0
Jinja2==3.1.6
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
keystoneauth1==5.11.1
keystonemiddleware==10.11.0
kombu==5.5.4
Mako==1.3.10
MarkupSafe==3.0.2
mistral==20.0.0
mistral-lib==3.4.0
mistralai==1.9.2
msgpack==1.1.1
netaddr==1.3.0
networkx==3.5
os-service-types==1.8.0
oslo.cache==3.11.0
oslo.concurrency==7.1.0
oslo.config==10.0.0
oslo.context==6.0.0
oslo.db==17.3.0
oslo.i18n==6.5.1
oslo.log==7.2.0
oslo.messaging==17.0.0
oslo.metrics==0.12.0
oslo.middleware==6.6.0
oslo.policy==4.5.1
oslo.serialization==5.7.0
oslo.service==4.2.2
oslo.utils==9.0.0
osprofiler==4.3.0
packaging==25.0
paramiko==3.5.1
Paste==3.10.1
PasteDeploy==3.1.0
pbr==6.1.1
pecan==1.7.0
ply==3.11
prettytable==3.16.0
prometheus_client==0.22.1
psutil==7.0.0
pycadf==4.0.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
PyNaCl==1.5.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-keystoneclient==5.6.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
repoze.lru==0.7
requests==2.32.4
rfc3986==2.0.0
Routes==2.5.1
rpds-py==0.26.0
setuptools==80.9.0
simplegeneric==0.8.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
statsd==4.0.1
stevedore==5.4.1
tenacity==9.1.2
testresources==2.0.2
testscenarios==0.5.0
testtools==2.7.2
tooz==7.0.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
voluptuous==0.15.2
wcwidth==0.2.13
WebOb==1.8.9
wikipedia==1.4.0
wrapt==1.17.2
WSME==0.12.1
yappi==1.6.10
yaql==3.1.0
zipp==3.23.0
